Fix oc_storageclass function needs_update
There was a bug in oc_storageclass that would cause the module to fail
with default options when mount_options was set to None. The module
would try to cast None to a set, which would cause an exception. Add a
"or []" to the call to set() that makes it so that there is no cast of
None anymore.